// frontend/TransportesOrellanaSpa.Web/src/app/core/models/crear-camion.ts
export interface CrearCamion {

  patente: string;

  marca: string;

  modelo: string;

  ano: number;

  tipo: string;

  color: string;

  capacidad: string;

  motor: string;

  caballos: string;

  cilindrada: string;

  transmision: string;

  fechaRevisionTecnica: string;

  fechaPermisoCirculacion: string;

  fechaSeguroObligatorio: string;

  revisionAlDia: boolean;

  permisoAlDia: boolean;

  seguroAlDia: boolean;

}